SearchSlice
Chopped my search time-wise using coroutines and now I'm getting a neat visualization of the game states being searched (with a 5x slowdown in search speed). I wonder how much this can be optimized. (And eventually refactored to allow the player to play with the solver working in the background.)